[#37]refactor : post 컴포넌트 외부 호출에서  post page 내부호출로변경
